chore(organization): move change_owner from account to organization service
Ownership transfer is semantically an organization concern, not an
account concern. Follow-up to #11810.
